Move reflect based late init implementation to CachePolicy
This is now the only controller that consumes it.
