+++ pkg/controller/cloudfront/cachepolicy/lateinit_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cloudfront
+package cachepolicy
 
 import (
 	"reflect"
